refactor(project): remove theme-preval.js
