+++ babel.config.js
@@ -6,7 +6,7 @@ function replacementPlugin(env) {
 
 const sharedPlugins = [
   'macros',
-  'preval',
+  // 'preval',
   'dev-expression',
   'add-react-displayname',
   'babel-plugin-styled-components',
@@ -23,13 +23,13 @@ module.exports = {
     development: {
       presets: makePresets(process.env.BABEL_MODULE || false),
       plugins: [
-        ...(process.env.BABEL_MODULE === 'commonjs'
-          ? ['@babel/plugin-transform-modules-commonjs']
-          : process.env.STORYBOOK
-          ? []
-          : [
-              'transform-commonjs', // theme-preval is commonjs and needs to be transformed to esm
-            ]),
+        // ...(process.env.BABEL_MODULE === 'commonjs'
+        // ? ['@babel/plugin-transform-modules-commonjs']
+        // : process.env.STORYBOOK
+        // ? []
+        // : [
+        // 'transform-commonjs', // theme-preval is commonjs and needs to be transformed to esm
+        // ]),
         ...sharedPlugins,
         replacementPlugin('development'),
       ],
